Conflicts:
	usr/local/www/interfaces_lagg_edit.php
	usr/local/www/interfaces_ppps_edit.php
	usr/local/www/interfaces_wireless_edit.php

+++ usr/local/www/system_firmware_restorefullbackup.php
@@ -3,6 +3,7 @@
 /*
 	system_firmware_restorefullbackup.php
 	Copyright (C) 2011 Scott Ullrich
+	Copyright (C) 2013-2015 Electric Sheep Fencing, LP
 	All rights reserved.
 
 	originally part of m0n0wall (http://m0n0.ch/wall)
@@ -52,16 +53,18 @@ require("guiconfig.inc");
 require_once("filter.inc");
 require_once("shaper.inc");
 
-if($_POST['overwriteconfigxml'])
+if ($_POST['overwriteconfigxml']) {
 	touch("/tmp/do_not_restore_config.xml");
+}
 
-if($_GET['backupnow'])
+if ($_GET['backupnow']) {
 	mwexec_bg("/etc/rc.create_full_backup");
+}
 
-if($_GET['downloadbackup']) {
+if ($_GET['downloadbackup']) {
 	$filename = basename($_GET['downloadbackup']);
 	$path = "/root/{$filename}";
-	if(file_exists($path)) {
+	if (file_exists($path)) {
 		session_write_close();
 		ob_end_clean();
 		session_cache_limiter('public');
@@ -70,11 +73,11 @@ if($_GET['downloadbackup']) {
 		header("Cache-Control: ");
 		header("Pragma: ");
 		header("Content-Type: application/octet-stream");
-		header("Content-Length: " .(string)(filesize($path)) );
+		header("Content-Length: " .(string)(filesize($path)));
 		header('Content-Disposition: attachment; filename="'.$filename.'"');
 		header("Content-Transfer-Encoding: binary\n");
-		if($file = fopen("/root/{$filename}", 'rb')){
-			while( (!feof($file)) && (connection_status()==0) ){
+		if ($file = fopen("/root/{$filename}", 'rb')) {
+			while ((!feof($file)) && (connection_status() == 0)) {
 				print(fread($file, 1024*8));
 				flush();
 			}
@@ -86,22 +89,26 @@ if($_GET['downloadbackup']) {
 }
 
 if ($_GET['deletefile']) {
-	$filename = $_GET['deletefile'];
-	if(file_exists("/root/{$filename}")) {
+	$filename = basename($_GET['deletefile']);
+	if (file_exists("/root/{$filename}") && (preg_match("/pfSense-full-backup-\d+-\d+\.tgz/", $filename) == 1)) {
 		unlink("/root/" . $filename);
-		$savemsg = gettext("$filename has been deleted.");
+		$savemsg = htmlspecialchars($filename) . " " . gettext("has been deleted.");
+	} else {
+		$savemsg = htmlspecialchars($filename) . " " . gettext("has not been been deleted (invalid backup file or file does not exist).");
 	}
 }
 
 if ($_POST['restorefile']) {
-	$filename = $_POST['restorefile'];
-	if(file_exists("/root/{$filename}")) {
+	$filename = basename($_POST['restorefile']);
+	if (file_exists("/root/{$filename}") && (preg_match("/pfSense-full-backup-\d+-\d+\.tgz/", $filename) == 1)) {
 		mwexec_bg("/etc/rc.restore_full_backup /root/" . escapeshellcmd($filename));
-		$savemsg = gettext("The firewall is currently restoring $filename");
+		$savemsg = gettext("The firewall is currently restoring") . " " . htmlspecialchars($filename);
+	} else {
+		$savemsg = htmlspecialchars($filename) . " " . gettext("has not been been restored (invalid backup file or file does not exist).");
 	}
 }
 
-$pgtitle = array(gettext("Diagnostics"),gettext("Restore full backup"));
+$pgtitle = array(gettext("Diagnostics"), gettext("Restore full backup"));
 include("head.inc");
 
 ?>
@@ -110,11 +117,13 @@ include("head.inc");
 <?php include("fbegin.inc"); ?>
 <?php if ($input_errors) print_input_errors($input_errors); ?>
 <?php if ($savemsg) print_info_box($savemsg); ?>
-<?php if (is_subsystem_dirty('restore')): ?><p>
-<form action="reboot.php" method="post">
-<input name="Submit" type="hidden" value="Yes" />
-<?php print_info_box(gettext("The firewall configuration has been changed.") . "<br />" . gettext("The firewall is now rebooting."));?><br />
-</form>
+<?php if (is_subsystem_dirty('restore')): ?>
+<p>
+	<form action="reboot.php" method="post">
+		<input name="Submit" type="hidden" value="Yes" />
+		<?php print_info_box(gettext("The firewall configuration has been changed.") . "<br />" . gettext("The firewall is now rebooting."));?><br />
+	</form>
+</p>
 <?php endif; ?>
 <form action="system_firmware_restorefullbackup.php" method="post">
 <table width="100%" border="0" cellspacing="0" cellpadding="0" summary="restore full backup">
@@ -125,8 +134,9 @@ include("head.inc");
 	$tab_array[] = array(gettext("Manual Update"), false, "system_firmware.php");
 	$tab_array[] = array(gettext("Auto Update"), false, "system_firmware_check.php");
 	$tab_array[] = array(gettext("Updater Settings"), false, "system_firmware_settings.php");
-	if($g['hidedownloadbackup'] == false)
+	if ($g['hidedownloadbackup'] == false) {
 		$tab_array[] = array(gettext("Restore Full Backup"), true, "system_firmware_restorefullbackup.php");
+	}
 	display_top_tabs($tab_array);
 ?>
 		</td>
@@ -145,20 +155,20 @@ include("head.inc");
 				chdir("/root");
 				$available_restore_files = glob("pfSense-full-backup-*");
 				$counter = 0;
-				foreach($available_restore_files as $arf) {
+				foreach ($available_restore_files as $arf) {
 					$counter++;
 					$size = exec("gzip -l /root/$arf | grep -v compressed | awk '{ print $2 }'");
 					echo "<tr>";
-					echo "<td  class='listlr' width='50%' colspan='1'>";
+					echo "<td class='listlr' width='50%' colspan='1'>";
 					echo "<input type='radio' name='restorefile' value='$arf' /> $arf";
 					echo "</td>";
-					echo "<td  class='listr' width='30%' colspan='1'>";
+					echo "<td class='listr' width='30%' colspan='1'>";
 					echo date ("F d Y H:i:s", filemtime($arf));
 					echo "</td>";
-					echo "<td  class='listr' width='40%' colspan='1'>";
+					echo "<td class='listr' width='40%' colspan='1'>";
 					echo format_bytes($size);
 					echo "</td>";
-					echo "<td  class='listr nowrap' width='20%' colspan='1'>";
+					echo "<td class='listr nowrap' width='20%' colspan='1'>";
 					echo "<a onclick=\"return confirm('" . gettext("Do you really want to delete this backup?") . "')\" href='system_firmware_restorefullbackup.php?deletefile=" . htmlspecialchars($arf) . "'>";
 					echo gettext("Delete");
 					echo "</a> | ";
@@ -168,9 +178,9 @@ include("head.inc");
 					echo "</td>";
 					echo "</tr>";
 				}
-				if($counter == 0) {
+				if ($counter == 0) {
 					echo "<tr>";
-					echo "<td  class='listlr' width='100%' colspan='4' align='center'>";
+					echo "<td class='listlr' width='100%' colspan='4' align='center'>";
 					echo gettext("Could not locate any previous backups.");
 					echo "</td>";
 					echo "</tr>";
@@ -203,7 +213,8 @@ decrypt_change();
 </html>
 <?php
 
-if (is_subsystem_dirty('restore'))
+if (is_subsystem_dirty('restore')) {
 	system_reboot();
+}
 
 ?>
